// MercaditoTec/src/app/models/course-tutored-interface.ts
export interface CourseTutoredInterface {
    idCursoTutorado: number;
    idTutorado: number;
    nombreTutorado: string;
    idCursoTutor: number;
    idTutor: number;
    nombreTutor: string;
    calificacionPromedioTutor: number;
    idCurso: number;
    nombreCurso: string;
    codigo: string;
    notaObtenida: number;
    temas: string;
}